Split out constants, and prep for future GUI work
