+++ KotorMessageInjector/KotorHelpers.cs
@@ -9,658 +9,8 @@ namespace KotorMessageInjector
         public KotorVersionNotFoundException(string message) : base(message) { }
     }
 
-    public static class KotorHelpers
+    public static partial class KotorHelpers
     {
-        private const uint KOTOR_OFFSET_SERVER_DEBUG_MODE = 0x1006c;
-        private const uint KOTOR_OFFSET_CLIENT_PLAYER_ID = 0x20;
-        private const uint KOTOR_OFFSET_CLIENT = 0x4;
-        private const uint KOTOR_OFFSET_SERVER = 0x8;
-        private const uint KOTOR_OFFSET_INTERNAL = 0x4;
-        private const uint KOTOR_OFFSET_LOADBAR = 0x278;
-        private const uint KOTOR_OFFSET_LAST_TARGET = 0x2B4;
-        private const uint KOTOR_OFFSET_MODULE = 0x18;
-        private const uint KOTOR_OFFSET_AREA = 0x48;
-        private const uint KOTOR_OFFSET_FACTION_MANAGER = 0x10054;
-        private const uint KOTOR_OFFSET_CLIENT_ANIM_BASE = 0x68;
-        private const uint KOTOR_OFFSET_ANIM_BASE_GOB = 0xb8;
-
-        private static IntPtr KOTOR_1_APPMANAGER = (IntPtr)0x007a39fc;
-        private static IntPtr KOTOR_1_DEACTIVATE_RENDER_WINDOW = (IntPtr)0x00401d90;
-        private const uint KOTOR_1_GOG_MODULE_SIZE = 4640768;
-        private const uint KOTOR_1_STEAM_MODULE_SIZE = 4993024;
-        private const uint KOTOR_1_LOAD_DIRECTION = 0xc8;
-        private const uint KOTOR_1_OFFSET_PARTY_TABLE = 0x1b770;
-        private const uint KOTOR_1_OFFSET_CHEAT_USED = 0x194;
-        private const uint KOTOR_1_OFFSET_SCENE = 0x184;
-        private const uint KOTOR_1_OFFSET_CREATURE_STATS = 0xa74;
-        private const uint KOTOR_1_OFFSET_CREATURE_STATS_RUNRATE = 0x198;
-
-        private static IntPtr KOTOR_2_APPMANAGER = (IntPtr)0x00a11c04;
-        private static IntPtr KOTOR_2_STEAM_APPMANAGER = (IntPtr)0x00a1b4a4;
-        private const uint KOTOR_2_STEAM_MODULE_SIZE = 7049216;
-        private const uint KOTOR_2_GOG_MODULE_SIZE = 7012352;
-        private const uint KOTOR_2_LOAD_DIRECTION = 0xd0;
-        private const uint KOTOR_2_GOG_OFFSET_PARTY_TABLE = 0x1f0b4;
-        private const uint KOTOR_2_OFFSET_CHEAT_USED = 0x23c;
-        private const uint KOTOR_2_OFFSET_SCENE = 0x188;
-        private const uint KOTOR_2_OFFSET_CREATURE_STATS = 0x1198;
-        private const uint KOTOR_2_OFFSET_CREATURE_STATS_RUNRATE = 0x1a8;
-
-        [Flags]
-        public enum CLIENT_OBJECT_UPDATE_FLAGS : uint
-        {
-            POSITION            = 0b00000000000000000000000000000001,
-            ORIENTATION         = 0b00000000000000000000000000000010,
-            ANIMATION           = 0b00000000000000000000000000000100,
-            VFX                 = 0b00000000000000000000000000001000,
-            OBJECT_INTERACTION  = 0b00000000000000000000000000010000, //Specific to Doors, Placeables, and Triggers
-            PORTRAIT            = 0b00000000000000000000000000100000,
-            // Below are specific to Creatures
-            // TODO: Fill these out
-
-        }
-
-        public enum GAME_OBJECT_TYPES : byte
-        {
-            OBJECT_0      = 0,
-            OBJECT_1      = 1,
-            OBJECT_2      = 2,
-            MODULE        = 3,
-            AREA          = 4,
-            CREATURE      = 5,
-            ITEM          = 6,
-            TRIGGER       = 7,
-            PROJECTILE    = 8,
-            PLACEABLE     = 9,
-            DOOR          = 10,
-            AREAOFEFFECT  = 11,
-            WAYPOINT      = 12,
-            ENCOUNTER     = 13,
-            STORE         = 14,
-            OBJECT_f      = 15,
-            SOUND         = 16,
-        }
-
-        public enum ATTRIBUTES
-        {
-            STR,
-            DEX,
-            CON,
-            INT,
-            WIS,
-            CHA
-        }
-
-        public enum SKILLS : byte
-        {
-            COMPUTER_USE = 0,
-            DEMOLITIONS = 1,
-            STEALTH = 2,
-            AWARNESS = 3,
-            PERSUADE = 4,
-            REPAIR = 5,
-            SECURITY = 6,
-            TREAT_INJURY = 7,
-        }
-
-        public enum FEATS : ushort
-        {
-            UNDEFINED = 0,
-            ADVANCED_JEDI_DEFENSE = 1,
-            XXXX_ADVANCED_GUARD_STANCE = 2,
-            TWO_WEAPON_FIGHTING = 3,
-            ARMOUR_PROF_HEAVY = 4,
-            ARMOUR_PROF_LIGHT = 5,
-            ARMOUR_PROF_MEDIUM = 6,
-            CAUTIOUS = 7,
-            CRITICAL_STRIKE = 8,
-            TWO_WEAPON_ADVANCED = 9,
-            EMPATHY = 10,
-            FLURRY = 11,
-            GEAR_HEAD = 12,
-            CONDITIONING = 13,
-            IMPLANT_LEVEL_1 = 14,
-            IMPLANT_LEVEL_2 = 15,
-            IMPLANT_LEVEL_3 = 16,
-            IMPROVED_POWER_ATTACK = 17,
-            IMPROVED_POWER_BLAST = 18,
-            IMPROVED_CRITICAL_STRIKE = 19,
-            IMPROVED_SNIPER_SHOT = 20,
-            IMPROVED_CONDITIONING = 21,
-            MASTER_CONDITIONING = 22,
-            MASTER_JEDI_DEFENSE = 24,
-            XXXX_MASTER_GUARD_STANCE = 25,
-            MULTI_SHOT = 26,
-            XXXX_PERCEPTIVE = 27,
-            POWER_ATTACK = 28,
-            POWER_BLAST = 29,
-            RAPID_SHOT = 30,
-            SNIPER_SHOT = 31,
-            WEAPON_FOCUS_BLASTER = 32,
-            WEAPON_FOCUS_BLASTER_RIFLE = 33,
-            XXXX_WEAPON_FOCUS_GRENADE = 34,
-            WEAPON_FOCUS_HEAVY_WEAPONS = 35,
-            WEAPON_FOCUS_LIGHTSABER = 36,
-            WEAPON_FOCUS_MELEE_WEAPONS = 37,
-            XXXX_WEAPON_FOCUS_SIMPLE_WEAPONS = 38,
-            WEAPON_PROF_BLASTER = 39,
-            WEAPON_PROF_BLASTER_RIFLE = 40,
-            XXXX_WEAPON_PROF_GRENADE = 41,
-            WEAPON_PROF_HEAVY_WEAPONS = 42,
-            WEAPON_PROF_LIGHTSABER = 43,
-            WEAPON_PROF_MELEE_WEAPONS = 44,
-            XXXX_WEAPON_PROF_SIMPLE_WEAPONS = 45,
-            WEAPON_SPEC_BLASTER = 46,
-            WEAPON_SPEC_BLASTER_RIFLE = 47,
-            XXXX_WEAPON_SPEC_GRENADE = 48,
-            WEAPON_SPEC_HEAVY_WEAPONS = 49,
-            WEAPON_SPEC_LIGHTSABER = 50,
-            WEAPON_SPEC_MELEE_WEAPONS = 51,
-            XXXX_WEAPON_SPEC_SIMPLE_WEAPONS = 52,
-            WHIRLWIND_ATTACK = 53,
-            XXXX_GUARD_STANCE = 54,
-            JEDI_DEFENSE = 55,
-            UNCANNY_DODGE_1 = 56,
-            UNCANNY_DODGE_2 = 57,
-            XXXX_SKILL_FOCUS_COMPUTER_USE = 58,
-            SNEAK_ATTACK_1D6 = 60,
-            SNEAK_ATTACK_2D6 = 61,
-            SNEAK_ATTACK_3D6 = 62,
-            SNEAK_ATTACK_4D6 = 63,
-            SNEAK_ATTACK_5D6 = 64,
-            SNEAK_ATTACK_6D6 = 65,
-            SNEAK_ATTACK_7D6 = 66,
-            SNEAK_ATTACK_8D6 = 67,
-            SNEAK_ATTACK_9D6 = 68,
-            SNEAK_ATTACK_10D6 = 69,
-            XXXX_SKILL_FOCUS_DEMOLITIONS = 70,
-            XXXX_SKILL_FOCUS_STEALTH = 71,
-            XXXX_SKILL_FOCUS_AWARENESS = 72,
-            XXXX_SKILL_FOCUS_PERSUADE = 73,
-            XXXX_SKILL_FOCUS_REPAIR = 74,
-            XXXX_SKILL_FOCUS_SECURITY = 75,
-            XXXX_SKILL_FOCUS_TREAT_INJURY = 76,
-            MASTER_SNIPER_SHOT = 77,
-            DROID_UPGRADE_1 = 78,
-            DROID_UPGRADE_2 = 79,
-            DROID_UPGRADE_3 = 80,
-            MASTER_CRITICAL_STRIKE = 81,
-            MASTER_POWER_BLAST = 82,
-            MASTER_POWER_ATTACK = 83,
-            TOUGHNESS = 84,
-            TWO_WEAPON_MASTERY = 85,
-            XXXX_FORCE_FOCUS_ALTER = 86,
-            XXXX_FORCE_FOCUS_CONTROL = 87,
-            FORCE_FOCUS = 88,
-            FORCE_FOCUS_ADVANCED = 89,
-            FORCE_FOCUS_MASTERY = 90,
-            IMPROVED_FLURRY = 91,
-            IMPROVED_RAPID_SHOT = 92,
-            PROFICIENCY_ALL = 93,
-            BATTLE_MEDITATION = 94,
-            WOOKIE_ENDURANCE = 95,
-            BLASTER_INTEGRATION = 96,
-            FORCE_CAMOFLAGE = 97,
-            FORCE_IMMUNITY_FEAR = 98,
-            FORCE_IMMUNITY_STUN = 99,
-            FORCE_IMMUNITY_PARALYSIS = 100,
-            FORCE_JUMP = 101,
-            FORCE_JUMP_ADVANCED = 102,
-            FORCE_JUMP_MASTERY = 103,
-            SCOUNDRELS_LUCK = 104,
-            IMPROVED_SCOUNDRELS_LUCK = 105,
-            MASTER_SCOUNDRELS_LUCK = 106,
-            JEDI_SENSE = 107,
-            KNIGHT_SENSE = 108,
-            MASTER_SENSE = 109,
-            LOGIC_UPGRADE_COMBAT = 110,
-            LOGIC_UPGRADE_TACTICIAN = 111,
-            LOGIC_UPGRADE_BATTLE_DROID = 112,
-            DUELING = 113,
-            ADVANCED_DUELING = 114,
-            MASTER_DUELING = 115,
-            FORCE_SENSITIVE = 116,
-            IMPROVED_CAUTION = 117,
-            MASTER_CAUTION = 118,
-            GEAR_HEAD_ADEPT = 119,
-            GEAR_HEAD_MASTER = 120,
-            IMPROVED_EMPATHY = 121,
-            MASTER_EMPATHY = 122,
-            IMPROVED_TOUGHNESS = 123,
-            MASTER_TOUGHNESS = 124,
-            //KOTOR 2 EXCLUSIVE FEATS
-            EVASION = 125,
-            TARGETING_1 = 126,
-            TARGETING_2 = 127,
-            TARGETING_3 = 128,
-            TARGETING_4 = 129,
-            TARGETING_5 = 130,
-            TARGETING_6 = 131,
-            TARGETING_7 = 132,
-            TARGETING_8 = 133,
-            XXXX_TARGETING_9 = 134,
-            XXXX_TARGETING_10 = 135,
-            XXXPRECISE_SHOT_I = 136,
-            XXXPRECISE_SHOT_II = 137,
-            XXXPRECISE_SHOT_III = 138,
-            CLOSE_COMBAT = 139,
-            IMPROVED_CLOSE_COMBAT = 140,
-            XXXIMPROVED_FORCE_CAMOUFLAGE = 141,
-            XXXMASTER_FORCE_CAMOUFLAGE = 142,
-            REGENERATE_FORCE_POINTS = 143,
-            XXXX_CRUSH_OPPOSITION_I = 144,
-            XXXX_CRUSH_OPPOSITION_II = 145,
-            XXXX_CRUSH_OPPOSITION_III = 146,
-            XXXX_CRUSH_OPPOSITION_IV = 147,
-            XXXX_CRUSH_OPPOSITION_V = 148,
-            DARK_SIDE_CORRUPTION = 149,
-            IGNORE_PAIN_I = 150,
-            IGNORE_PAIN_II = 151,
-            IGNORE_PAIN_III = 152,
-            INCREASE_COMBAT_DAMAGE_I = 153,
-            INCREASE_COMBAT_DAMAGE_II = 154,
-            INCREASE_COMBAT_DAMAGE_III = 155,
-            SUPER_WEAPON_FOCUS_LIGHTSABER_I = 156,
-            SUPER_WEAPON_FOCUS_LIGHTSABER_II = 157,
-            SUPER_WEAPON_FOCUS_LIGHTSABER_III = 158,
-            SUPER_WEAPON_FOCUS_2_WEAPON_I = 159,
-            SUPER_WEAPON_FOCUS_2_WEAPON_II = 160,
-            SUPER_WEAPON_FOCUS_2_WEAPON_III = 161,
-            XXXX_INSPIRE_FOLLOWERS_I = 162,
-            XXXX_INSPIRE_FOLLOWERS_II = 163,
-            XXXX_INSPIRE_FOLLOWERS_III = 164,
-            XXXX_INSPIRE_FOLLOWERS_IV = 165,
-            XXXX_INSPIRE_FOLLOWERS_V = 166,
-            LIGHT_SIDE_ENLIGHTENMENT = 167,
-            DEFLECT = 168,
-            INNER_STRENGTH_I = 169,
-            INNER_STRENGTH_II = 170,
-            INNER_STRENGTH_III = 171,
-            INCREASE_MELEE_DAMAGE_I = 172,
-            INCREASE_MELEE_DAMAGE_II = 173,
-            INCREASE_MELEE_DAMAGE_III = 174,
-            CRAFT = 175,
-            MASTERCRAFT_WEAPONS_I = 176,
-            MASTERCRAFT_WEAPONS_II = 177,
-            MASTERCRAFT_WEAPONS_III = 178,
-            MASTERCRAFT_ARMOR_I = 179,
-            MASTERCRAFT_ARMOR_II = 180,
-            MASTERCRAFT_ARMOR_III = 181,
-            DROID_INTERFACE = 182,
-            CLASS_SKILL_AWARENESS = 183,
-            CLASS_SKILL_COMPUTER_USE = 184,
-            CLASS_SKILL_DEMOLITIONS = 185,
-            CLASS_SKILL_REPAIR = 186,
-            CLASS_SKILL_SECURITY = 187,
-            CLASS_SKILL_STEALTH = 188,
-            CLASS_SKILL_TREAT_INJURY = 189,
-            DUAL_STRIKE = 190,
-            IMPROVED_DUAL_STRIKE = 191,
-            MASTER_DUAL_STRIKE = 192,
-            FINESSE_LIGHTSABERS = 193,
-            FINESSE_MELEE_WEAPONS = 194,
-            XXXX_MOBILITY = 195,
-            REGENERATE_VITALITY_POINTS = 196,
-            STEALTH_RUN = 197,
-            KINETIC_COMBAT = 198,
-            SURVIVAL = 199,
-            MANDALORIAN_COURAGE = 200,
-            PERSONAL_CLOAKING_SHIELD = 201,
-            MENTOR = 202,
-            IMPLANT_SWITCHING = 203,
-            SPIRIT = 204,
-            FORCE_CHAIN = 205,
-            WAR_VETERAN = 206,
-            COMPLEX_UNARMED_ANIMS = 207,
-            WEAPON_PROF_WRIST_MOUNTED = 208,
-            ECHANI_STRIKE_I = 209,
-            ECHANI_STRIKE_II = 210,
-            ECHANI_STRIKE_III = 211,
-            UNARMED_SPECIALIST_I = 212,
-            UNARMED_SPECIALIST_II = 213,
-            UNARMED_SPECIALIST_III = 214,
-            UNARMED_SPECIALIST_IV = 215,
-            UNARMED_SPECIALIST_V = 216,
-            UNARMED_SPECIALIST_VI = 217,
-            UNARMED_SPECIALIST_VII = 218,
-            UNARMED_SPECIALIST_VIII = 219,
-            SHIELD_BREAKER = 220,
-            REPULSOR_STRIKE = 221,
-            ELECTRICAL_STRIKE = 222,
-            GRAVITONIC_STRIKE = 223,
-            WOOKIEE_TOUGHNESS_II = 224,
-            WOOKIEE_TOUGHNESS_III = 225,
-            XXXPRECISE_SHOT_IV = 226,
-            XXXPRECISE_SHOT_V = 227,
-            ASSASSIN_PROTOCOL_I = 228,
-            ASSASSIN_PROTOCOL_II = 229,
-            ASSASSIN_PROTOCOL_III = 230,
-            WOOKIEE_RAGE_I = 231,
-            WOOKIEE_RAGE_II = 232,
-            WOOKIEE_RAGE_III = 233,
-            DROID_TRICK = 234,
-            DROID_CONFUSION = 235,
-            FIGHTING_SPIRIT = 236,
-            HEROIC_RESOLVE = 237,
-            MINE_IMMUNITY = 238,
-            POINT_GUARD = 239,
-            PRECISE_SHOT_I = 240,
-            PRECISE_SHOT_II = 241,
-            PRECISE_SHOT_III = 242,
-            PRECISE_SHOT_IV = 243,
-            PRECISE_SHOT_V = 244,
-
-        }
-
-        public enum CLASSES : byte
-        {
-            SOLDIER = 0,
-            SCOUT = 1,
-            SCOUNDREL = 2,
-            JEDI_GUARDIAN = 3,
-            JEDI_CONSULAR = 4,
-            JEDI_SENTINEL = 5,
-            COMBAT_DROID = 6,
-            EXPERT_DROID = 7,
-            MINION = 8,
-            // KotOR 2 Exclusive Classes
-            TECH_SPECIALIST = 9,
-            BOUNTY_HUNTER = 10,
-            JEDI_WEAPONMASTER = 11,
-            JEDI_MASTER = 12,
-            JEDI_WATCHMAN = 13,
-            SITH_MARAUDER = 14,
-            SITH_LORD = 15,
-            SITH_ASSASSIN = 16
-        }
-        
-        public enum SPELLS : int
-        {
-            FORCE_POWER_MASTER_ALTER_XXX = 0,
-            FORCE_POWER_MASTER_CONTROL_XXX = 1,
-            FORCE_POWER_MASTER_SENSE_XXX = 2,
-            FORCE_POWER_FORCE_JUMP_ADVANCED_XXX = 3,
-            FORCE_POWER_LIGHT_SABER_THROW_ADVANCED = 4,
-            FORCE_POWER_REGNERATION_ADVANCED_XXX = 5,
-            FORCE_POWER_AFFECT_MIND = 6,
-            FORCE_POWER_AFFLICTION = 7,
-            FORCE_POWER_SPEED_BURST = 8,
-            FORCE_POWER_CHOKE = 9,
-            FORCE_POWER_CURE = 10,
-            FORCE_POWER_DEATH_FIELD = 11,
-            FORCE_POWER_DROID_DISABLE = 12,
-            FORCE_POWER_DROID_DESTROY = 13,
-            FORCE_POWER_DOMINATE = 14,
-            FORCE_POWER_DRAIN_LIFE = 15,
-            FORCE_POWER_FEAR = 16,
-            FORCE_POWER_FORCE_ARMOR = 17,
-            FORCE_POWER_FORCE_AURA = 18,
-            FORCE_POWER_FORCE_BREACH = 19,
-            FORCE_POWER_FORCE_IMMUNITY = 20,
-            FORCE_POWER_FORCE_JUMP_XXX = 21,
-            FORCE_POWER_FORCE_VALOR = 22,
-            FORCE_POWER_FORCE_PUSH = 23,
-            FORCE_POWER_FORCE_SHIELD = 24,
-            FORCE_POWER_FORCE_STORM = 25,
-            FORCE_POWER_FORCE_WAVE = 26,
-            FORCE_POWER_FORCE_WHIRLWIND = 27,
-            FORCE_POWER_HEAL = 28,
-            FORCE_POWER_STASIS = 29,
-            FORCE_POWER_HORROR = 30,
-            FORCE_POWER_INSANITY = 31,
-            FORCE_POWER_KILL = 32,
-            FORCE_POWER_KNIGHT_VALOR = 33,
-            FORCE_POWER_KNIGHT_SPEED = 34,
-            FORCE_POWER_LIGHTNING = 35,
-            FORCE_POWER_MASTER_VALOR = 36,
-            FORCE_POWER_SPEED_MASTERY = 37,
-            FORCE_POWER_PLAGUE = 38,
-            FORCE_POWER_REGENERATION_XXX = 39,
-            FORCE_POWER_RESIST_ENERGY_2 = 40,
-            FORCE_POWER_RESIST_FORCE = 41,
-            FORCE_POWER_RESIST_ENERGY_1 = 42,
-            FORCE_POWER_SHOCK = 43,
-            FORCE_POWER_MASS_STASIS = 44,
-            FORCE_POWER_SLOW = 45,
-            FORCE_POWER_STUN = 46,
-            FORCE_POWER_DROID_STUN = 47,
-            FORCE_POWER_SUPRESS_FORCE = 48,
-            FORCE_POWER_LIGHT_SABER_THROW = 49,
-            FORCE_POWER_WOUND = 50,
-            SPECIAL_ABILITY_BATTLE_MEDITATION_XXXX = 51,
-            SPECIAL_ABILITY_BODY_FUEL = 52,
-            SPECIAL_ABILITY_COMBAT_REGENERATION = 53,
-            SPECIAL_ABILITY_WARRIOR_STANCE_XXXX = 54,
-            SPECIAL_ABILITY_SENTINEL_STANCE_XXXX = 55,
-            SPECIAL_ABILITY_DOMINATE_MIND_XXX = 56,
-            SPECIAL_ABILITY_PSYCHIC_STANCE = 57,
-            SPECIAL_ABILITY_CATHAR_REFLEXES_XXX = 58,
-            SPECIAL_ABILITY_ENHANCED_SENSES = 59,
-            SPECIAL_ABILITY_CAMOFLAGE = 60,
-            SPECIAL_ABILITY_TAUNT_XXX = 61,
-            SPECIAL_ABILITY_WHIRLING_DERVISH_XXX = 62,
-            SPECIAL_ABILITY_RAGE = 63,
-            ITEM_ABILITY_MED_PACK = 64,
-            ITEM_ABILITY_ADVANCED_MED_PACK = 65,
-            ITEM_ABILITY_LIFE_SUPPORT_KIT = 66,
-            ITEM_ABILITY_REMOVE_POISON = 67,
-            PLOT_MAN_ABILITY_SONIC_UPGRADE = 68,
-            ITEM_ABILITY_ADRENAL_SHOT_STRENGTH = 69,
-            ITEM_ABILITY_ADRENAL_SHOT_ALACRITY = 70,
-            ITEM_ABILITY_ADRENAL_SHOT_STAMINA = 71,
-            ITEM_ABILITY_HYPER_ADRENAL_SHOT_STRENGTH = 72,
-            ITEM_ABILITY_HYPER_ADRENAL_SHOT_ALACRITY = 73,
-            ITEM_ABILITY_HYPER_ADRENAL_SHOT_STAMINA = 74,
-            ITEM_ABILITY_COMBAT_SHOT_1 = 75,
-            ITEM_ABILITY_COMBAT_SHOT_2 = 76,
-            ITEM_ABILITY_COMBAT_SHOT_3 = 77,
-            ITEM_ABILITY_SHIELD_DISRUPTER = 78,
-            ITEM_ABILITY_SHIELD_DISRUPTER_ADVANCED = 79,
-            ITEM_ABILITY_OIL_SLICK = 80,
-            ITEM_ABILITY_GRAVITY_GENERATOR = 81,
-            ITEM_ABILITY_GRAVITY_GENERATOR_ADVANCED = 82,
-            MONSTER_ABILITY_SLAM_ATTACK = 83,
-            ITEM_ABILITY_DROID_REPAIR_KIT = 84,
-            PLOT_TAT_SAND_PERSON_DISGUISE = 85,
-            PLOT_TAR_SITH_DISGUISE = 86,
-            ITEM_ABILITY_GRENADE_FRAGMENTATION = 87,
-            ITEM_ABILITY_GRENADE_STUN = 88,
-            ITEM_ABILITY_GRENADE_THERMAL_DETONATOR = 89,
-            ITEM_ABILITY_GRENADE_POISON = 90,
-            ITEM_ABILITY_GRENADE_SONIC = 91,
-            ITEM_ABILITY_GRENADE_ADHESIVE = 92,
-            ITEM_ABILITY_GRENADE_CRYOBAN = 93,
-            ITEM_ABILITY_GRENADE_FIRE = 94,
-            ITEM_ABILITY_GRENADE_ION = 95,
-            PLOT_UNK_GRENADE_RANCOR = 96,
-            PLOT_KOR_GRENADE_FIRE = 97,
-            PLOT_KOR_GRENADE_COLD = 98,
-            ITEM_ABILITY_ENERGY_SHIELD = 99,
-            ITEM_ABILITY_SITH_ENERGY_SHIELD = 100,
-            ITEM_ABILITY_ARKANIAN_ENERGY_SHIELD = 101,
-            ITEM_ABILITY_ECHANI_SHIELD = 102,
-            ITEM_ABILITY_MANDLORIAN_MELEE_SHIELD = 103,
-            ITEM_ABILITY_MANDLORIAN_POWER_SHIELD = 104,
-            ITEM_ABILITY_ECHANI_DUELING_SHIELD = 105,
-            ITEM_ABILITY_YUSANIS_DUELING_SHIELD = 106,
-            ITEM_ABILITY_VERPINE_ENERGY_SHIELD = 107,
-            ITEM_ABILITY_CLOAKING_DEVICE = 108,
-            PLOT_LEV_DAMAGE_SPACE_SUIT = 109,
-            DROID_ITEM_ENERGY_SHIELD_1 = 110,
-            DROID_ITEM_ENERGY_SHIELD_2 = 111,
-            DROID_ITEM_ENERGY_SHIELD_3 = 112,
-            DROID_ITEM_HAZARD_SHIELD_1 = 113,
-            DROID_ITEM_HAZARD_SHIELD_2 = 114,
-            DROID_ITEM_HAZARD_SHIELD_3 = 115,
-            DROID_ITEM_STUN_RAY_1 = 116,
-            DROID_ITEM_STUN_RAY_2 = 117,
-            DROID_ITEM_SHIELD_DISRUPTOR_1 = 118,
-            DROID_ITEM_SHIELD_DISRUPTOR_2 = 119,
-            DROID_ITEM_FLAME_THROWER_1 = 120,
-            DROID_ITEM_FLAME_THROWER_2 = 121,
-            DROID_ITEM_CARBONITE_PROJECTOR_1 = 122,
-            DROID_ITEM_CARBONITE_PROJECTOR_2 = 123,
-            DROID_ITEM_GRAVITY_GENERATOR_1 = 124,
-            DROID_ITEM_GRAVITY_GENERATOR_2 = 125,
-            PLOT_UNK_COLD_SERUM = 126,
-            ITEM_ABILITY_DROID_REPAIR_KIT_ADV = 127,
-            ITEM_ABILITY_DROID_CONSTRUCTION_KIT = 128,
-            ITEM_ABILITY_RECOVERY_STIM = 129,
-            MONSTER_ABILITY_FIRE_BREATH = 130,
-            // Kotor 2 Only Powers
-            MONSTER_ABILITY_SONIC_HOWL = 131,
-            ITEM_ABILITY_HEAT_SHIELD = 132,
-            FORCE_POWER_MASTER_ENERGY_RESISTANCE = 133,
-            FORCE_POWER_MASTER_HEAL = 134,
-            FORCE_POWER_FORCE_BARRIER = 135,
-            FORCE_POWER_IMPROVED_FORCE_BARRIER = 136,
-            FORCE_POWER_MASTER_FORCE_BARRIER = 137,
-            FORCE_POWER_BATTLE_MEDITATION_PC = 138,
-            FORCE_POWER_IMPROVED_BATTLE_MEDITATION_PC = 139,
-            FORCE_POWER_MASTER_BATTLE_MEDITATION_PC = 140,
-            FORCE_POWER_BAT_MED_ENEMY = 141,
-            FORCE_POWER_IMP_BAT_MED_ENEMY = 142,
-            FORCE_POWER_MAS_BAT_MED_ENEMY = 143,
-            FORCE_POWER_CRUSH_OPPOSITION_I = 144,
-            FORCE_POWER_CRUSH_OPPOSITION_II = 145,
-            FORCE_POWER_CRUSH_OPPOSITION_III = 146,
-            FORCE_POWER_CRUSH_OPPOSITION_IV = 147,
-            FORCE_POWER_CRUSH_OPPOSITION_V = 148,
-            FORCE_POWER_CRUSH_OPPOSITION_VI = 149,
-            FORCE_POWER_FORCE_BODY = 150,
-            FORCE_POWER_IMPROVED_FORCE_BODY = 151,
-            FORCE_POWER_MASTER_FORCE_BODY = 152,
-            FORCE_POWER_DRAIN_FORCE = 153,
-            FORCE_POWER_IMPROVED_DRAIN_FORCE = 154,
-            FORCE_POWER_MASTER_DRAIN_FORCE = 155,
-            FORCE_POWER_FORCE_CAMOUFLAGE = 156,
-            FORCE_POWER_IMPROVED_FORCE_CAMOUFLAGE = 157,
-            FORCE_POWER_MASTER_FORCE_CAMOUFLAGE = 158,
-            FORCE_POWER_FORCE_SCREAM = 159,
-            FORCE_POWER_IMPROVED_FORCE_SCREAM = 160,
-            FORCE_POWER_MASTER_FORCE_SCREAM = 161,
-            FORCE_POWER_FORCE_REPULSION = 162,
-            FORCE_POWER_FORCE_REDIRECTION = 163,
-            FORCE_POWER_FURY = 164,
-            FORCE_POWER_IMPROVED_FURY = 165,
-            FORCE_POWER_MASTER_FURY = 166,
-            FORCE_POWER_INSPIRE_FOLLOWERS_I = 167,
-            FORCE_POWER_INSPIRE_FOLLOWERS_II = 168,
-            FORCE_POWER_INSPIRE_FOLLOWERS_III = 169,
-            FORCE_POWER_INSPIRE_FOLLOWERS_IV = 170,
-            FORCE_POWER_INSPIRE_FOLLOWERS_V = 171,
-            FORCE_POWER_INSPIRE_FOLLOWERS_VI = 172,
-            FORCE_POWER_REVITALIZE = 173,
-            FORCE_POWER_IMPROVED_REVITALIZE = 174,
-            FORCE_POWER_MASTER_REVITALIZE = 175,
-            FORCE_POWER_FORCE_SIGHT = 176,
-            FORCE_POWER_FORCE_CRUSH = 177,
-            FORCE_POWER_PRECOGNITION = 178,
-            FORCE_POWER_BATTLE_PRECOGNITION = 179,
-            FORCE_POWER_FORCE_ENLIGHTENMENT = 180,
-            FORCE_POWER_MIND_TRICK = 181,
-            FORCE_POWER_BEAST_TRICK = 182,
-            XXXFORCE_POWER_IMPROVED_BEAST_CONTROL = 183,
-            FORCE_POWER_BEAST_CONFUSION = 184,
-            ITEM_ABILITY_ROCKET_EXPLOSIVE = 185,
-            ITEM_ABILITY_GRENADE_SONIC_2 = 186,
-            ITEM_ABILITY_GRENADE_SONIC_3 = 187,
-            DROID_ITEM_FLAME_THROWER_3 = 188,
-            DROID_ITEM_FLAME_THROWER_4 = 189,
-            ITEM_ABILITY_ROCKET_PIERCING = 190,
-            ITEM_ABILITY_ROCKET_POISON = 191,
-            ITEM_ABILITY_ROCKET_ION = 192,
-            ITEM_ABILITY_ROCKET_PLASMA = 193,
-            ITEM_ABILITY_ROCKET_CONCUSSION = 194,
-            ITEM_ABILITY_DART_POISON = 195,
-            ITEM_ABILITY_DART_KYBER = 196,
-            ITEM_ABILITY_DART_PARALYSIS = 197,
-            ITEM_ABILITY_DART_TRANQUILIZER = 198,
-            ITEM_ABILITY_DART_LAWN = 199,
-            FORCE_POWER_CONFUSION = 200,
-            FORCE_POWER_DROID_TRICK = 201,
-            XXX_FORCE_POWER_MEDITATION = 202,
-            XXX_FORCE_POWER_KNIGHT_MEDITATION = 203,
-            XXX_FORCE_POWER_MASTER_MEDITATION = 204,
-            XXX_FORM_LIGHTSABER_PADAWAN_I = 205,
-            XXX_FORM_LIGHTSABER_PADAWAN_II = 206,
-            XXX_FORM_LIGHTSABER_PADAWAN_III = 207,
-            XXX_FORM_LIGHTSABER_DAKLEAN_I = 208,
-            XXX_FORM_LIGHTSABER_DAKLEAN_II = 209,
-            XXX_FORM_LIGHTSABER_DAKLEAN_III = 210,
-            XXX_FORM_LIGHTSABER_SENTINEL_I = 211,
-            XXX_FORM_LIGHTSABER_SENTINEL_II = 212,
-            XXX_FORM_LIGHTSABER_SENTINEL_III = 213,
-            XXX_FORM_LIGHTSABER_SODAK_I = 214,
-            XXX_FORM_LIGHTSABER_SODAK_II = 215,
-            XXX_FORM_LIGHTSABER_SODAK_III = 216,
-            XXX_FORM_LIGHTSABER_ANCIENT_I = 217,
-            XXX_FORM_LIGHTSABER_ANCIENT_II = 218,
-            XXX_FORM_LIGHTSABER_ANCIENT_III = 219,
-            XXX_FORM_LIGHTSABER_MASTER_I = 220,
-            XXX_FORM_LIGHTSABER_MASTER_II = 221,
-            XXX_FORM_LIGHTSABER_MASTER_III = 222,
-            XXX_FORM_CONSULAR_FORCE_FOCUS_I = 223,
-            XXX_FORM_CONSULAR_FORCE_FOCUS_II = 224,
-            XXX_FORM_CONSULAR_FORCE_FOCUS_III = 225,
-            XXX_FORM_CONSULAR_ENDURING_FORCE_I = 226,
-            XXX_FORM_CONSULAR_ENDURING_FORCE_II = 227,
-            XXX_FORM_CONSULAR_ENDURING_FORCE_III = 228,
-            XXX_FORM_CONSULAR_FORCE_AMPLIFICATION_I = 229,
-            XXX_FORM_CONSULAR_FORCE_AMPLIFICATION_II = 230,
-            XXX_FORM_CONSULAR_FORCE_AMPLIFICATION_III = 231,
-            XXX_FORM_CONSULAR_FORCE_SHELL_I = 232,
-            XXX_FORM_CONSULAR_FORCE_SHELL_II = 233,
-            XXX_FORM_CONSULAR_FORCE_SHELL_III = 234,
-            XXX_FORM_CONSULAR_FORCE_POTENCY_I = 235,
-            XXX_FORM_CONSULAR_FORCE_POTENCY_II = 236,
-            XXX_FORM_CONSULAR_FORCE_POTENCY_III = 237,
-            XXX_FORM_CONSULAR_REGENERATION_I = 238,
-            XXX_FORM_CONSULAR_REGENERATION_II = 239,
-            XXX_FORM_CONSULAR_REGENERATION_III = 240,
-            XXX_FORM_CONSULAR_POWER_OF_THE_DARK_SIDE_I = 241,
-            XXX_FORM_CONSULAR_POWER_OF_THE_DARK_SIDE_II = 242,
-            XXX_FORM_CONSULAR_POWER_OF_THE_DARK_SIDE_III = 243,
-            DROID_ITEM_CARBONITE_PROJECTOR_3 = 244,
-            DROID_ITEM_CARBONITE_PROJECTOR_4 = 245,
-            ITEM_ABILITY_GRENADE_LAUNCHED_FRAGMENTATION = 246,
-            ITEM_ABILITY_GRENADE_LAUNCHED_STUN = 247,
-            ITEM_ABILITY_GRENADE_LAUNCHED_THERMAL_DETONATOR = 248,
-            ITEM_ABILITY_GRENADE_LAUNCHED_POISON = 249,
-            ITEM_ABILITY_GRENADE_LAUNCHED_SONIC = 250,
-            ITEM_ABILITY_GRENADE_LAUNCHED_ADHESIVE = 251,
-            ITEM_ABILITY_GRENADE_LAUNCHED_CRYOBAN = 252,
-            ITEM_ABILITY_GRENADE_LAUNCHED_FIRE = 253,
-            ITEM_ABILITY_GRENADE_LAUNCHED_ION = 254,
-            ITEM_ABILITY_GRENADE_LAUNCHED_SONIC_2 = 255,
-            ITEM_ABILITY_GRENADE_LAUNCHED_SONIC_3 = 256,
-            ITEM_ABILITY_DREXL_SHIELD = 257,
-            FORM_SABER_I_SHII_CHO = 258,
-            FORM_SABER_II_MAKASHI = 259,
-            FORM_SABER_III_SORESU = 260,
-            FORM_SABER_IV_ATARU = 261,
-            FORM_SABER_V_SHIEN = 262,
-            FORM_SABER_VI_NIMAN = 263,
-            FORM_SABER_VII_JUYO = 264,
-            FORM_FORCE_I_FOCUS = 265,
-            FORM_FORCE_II_POTENCY = 266,
-            FORM_FORCE_III_AFFINITY = 267,
-            FORM_FORCE_IV_MASTERY = 268,
-            FORCE_POWER_DROID_CONFUSION = 269,
-            FORCE_POWER_BREATH_CONTROL = 270,
-            WOOKIEE_RAGE_I = 271,
-            WOOKIEE_RAGE_II = 272,
-            WOOKIEE_RAGE_III = 273,
-            DROID_ITEM_ION_BLAST_I = 274,
-            DROID_ITEM_ION_BLAST_II = 275,
-            DROID_ITEM_ION_BLAST_III = 276,
-            DROID_ITEM_PLASMA_THROWER = 277,
-            DROID_ITEM_CHARGE_ARM = 278,
-            DROID_ITEM_TOXIN_EMITTER = 279,
-            DROID_ITEM_BIO_ASSAULT_SPRAY = 280,
-            DROID_ITEM_MULTI_SPECTRAL_EMITTER = 281
-        }
-        
         public static IntPtr getRunningKotor()
         {
             IntPtr pHandle = OpenProcessByName("swkotor.exe");
@@ -1025,6 +375,19 @@ namespace KotorMessageInjector
             var partyTable = getServerPartyTable(processHandle);
 
             WriteProcessMemory(processHandle, (IntPtr)(partyTable + offset), inBytes, 4, out _);
+        }
+
+        public static uint getInGameGui(IntPtr processHandle)
+        {
+            byte[] outBytes = new byte[4];
+
+            int version = getGameVersion(processHandle);
+            uint offset = KOTOR_OFFSET_GUI_IN_GAME;
+
+            uint clientInternal = getClientInternal(processHandle);
+
+            ReadProcessMemory(processHandle, (IntPtr)(clientInternal + offset), outBytes, 4, out _);
+            return BitConverter.ToUInt32(outBytes, 0);
         }
     }
 }
